fix(competencies-ecoe): change StudentCompetencyEntityOrm grade column type to float

+++ src/competencies-ecoe/infrastructure/persistence/typeorm/entities/student-competency.entity.orm.ts
@@ -16,7 +16,7 @@ export class StudentCompetencyEntityOrm {
     @JoinColumn({ name: 'competency_id' })
     competency: CompetencyEntityOrm;
 
-    @Column('numeric')
+    @Column('float')
     grade: number;
 
     @Column()
